+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## edge
 
+*   Fix parsing of escapes in single-strings ('foo\n'). Only ' and \
+    characters now get escaped in single strings.
+
 *   Sourcemaps revamp. Lexer now tracks column and line info for ever token to
     produce much more accurate sourcemaps. All method calls are now located on
     the correct source line, and multi-line xstrings are now split to generate